ENH: make parameters --use-ipython-kernel and --ignore-invalid-plans take ON/OFF values
